feature: add navigation list component
Creates a new NavigationListComponent. It is an specialization of the ActionListComponent in which items are links (`a` elements), and these are highlighted based on the `url` passed to the component during initialization (matching link `urls` will be highlighted by default).

If no URL is specified during initialization the component will use the URL of the page it was rendered in as default URL.