Add new flash & flash notification components
The FlashComponent is just a wrapper with a turbo ID to replace by new FlashComponents or to append new FlashNotification components via turbo streams. FlashNotificationComponents are equipped with a stimulus controller which triggers an animation that reveals and hides the component, removing it from the DOM once completed. Users can cancel the animation clicking on the close button, which cancels the ongoing animation and triggers a new one to hide the element.