chore: Adjusts `Attribute` logic to handle multi-value HTML attributes.
In a previous PR I defined the ClassNameHelper. This helper concatenates multiple values representing CSS classes into a single string with all class names which can be passed to an HTML element data attribute.

Now I am using the same helper in the `Attribute` class to handle other HTML attributes which, similar to the `class` attribute, can accept multiple values.

For example, a `Stimulus` `data-action` attribute can define multiple actions, each one concatenated in a single string and separated by spaces. For attributes like this, the new +Attribute+ class can be extended in order to extract the value in the same way we did with the `class` attribute.