Define Parameters concern for nested association parameters
Via accepts nested attributes we can create with a single payload a record with all its associations. In Rails, the payload for an association must be suffixed with '_attributes'.

Via the new WithNestedAssociation concern we can easily define nested association parameters without the need to manually define them with the respective suffix. And elegant solution to handle these particular parameters.