fix: icon size
Icons were not being displayed with the right size. The problem was due to the `inline_svg_tag` helper not being able to handle flattened data attributes (for example "data-size"). Instead, the method only accepts a single `data` attribute as `Hash`.

This fix reverts the flattening of data attributes made by the `Html::Attributes` model, ensuring the data is passed in a compatible format.