refactor: button component data-scoped classes replaced by inline classes
The previous approach of using the data attribute to set different button scheme and sizes bloatted the DOM with Tailwind classes.

We only want to switch the style of a button when it is enabled or disabled, thus it has no sense to base the style in these other data attributes.