fix: Event now persist datetimes in two separated files.
Range field was not persisting the timestampz, only the range of dates. Thus we were not able to display the exact times of the events.

I added two new timestamp fields to display the proper times. This makes now the schedule field redundant, and should be removed in a future PR.